+++ src/views/search/SearchResultsView.jsx
@@ -1,5 +1,5 @@
 // SearchResultsView.jsx
-import React from "react";
+import React, { useState } from "react";
 import { useLocation } from "react-router-dom";
 import SearchCamping from "../../components/search/SearchCamping";
 import SearchResults from "../../components/search/SearchResults";
@@ -7,18 +7,21 @@ import "../../assets/css/components/search/searchResultsView.css";
 
 const SearchResultsView = () => {
   const location = useLocation();
-  const searchResults = location.state?.searchResults || [];
+  
+  // Estado local para almacenar los resultados de búsqueda
+  const [searchResults, setSearchResults] = useState(location.state?.searchResults || []);
 
+  // Función para manejar una nueva búsqueda
   const handleSearch = (data) => {
     console.log("Nueva búsqueda:", data);
-    // Actualiza los resultados en SearchResults
+    setSearchResults(data); // Actualizamos el estado local con los nuevos resultados
   };
 
   return (
     <>
       <div className="search-results-view">
-        <SearchCamping onSearch={handleSearch} />
-        <SearchResults results={searchResults} />
+        <SearchCamping onSearch={handleSearch} /> {/* Pasamos la función handleSearch */}
+        <SearchResults results={searchResults} /> {/* Renderizamos los resultados actualizados */}
       </div>
     </>
   );
